fix: GatherList 컴포넌트단위로  나누기
