+++ frontend/src/components/body/gather/ListBtn.vue
@@ -1,0 +1,20 @@
+<template>
+    <div class="gatherbtn">
+        <b-button size="lg" variant="outline-primary">나의 모임 조회</b-button>
+        <b-button size="lg" variant="outline-primary">모임 생성</b-button>
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+    .gatherbtn{
+        display: flex;
+        justify-content: end;
+        margin-right: 3rem;
+        padding-bottom: 1rem;
+        gap: 1rem;
+    }
+</style>
