+++ frontend/src/components/body/gather/ListTitle.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        <h1>모임 목록 조회</h1>
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
